requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langchain-chroma
chromadb
python-dotenv
openai
tiktoken
unstructured[all-docs]
pypdf
pdfminer.six
Pillow